# Customer Management Application/customer-management/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Customer API
    version: 0.0.1
paths:
    /v1/customer:
        post:
            tags:
                - Customer
            operationId: Customer_CreateCustomer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateCustomerRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateCustomerReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/customer/{id}:
        get:
            tags:
                - Customer
            operationId: Customer_GetCustomer
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetCustomerReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        put:
            tags:
                - Customer
            operationId: Customer_UpdateCustomer
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateCustomerRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateCustomerReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - Customer
            operationId: Customer_DeleteCustomer
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteCustomerReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/customers:
        get:
            tags:
                - Customer
            operationId: Customer_ListCustomer
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListCustomersReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        Address:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                address:
                    type: string
                pincode:
                    type: string
                type:
                    type: string
        CreateCustomerReply:
            type: object
            properties:
                customer:
                    $ref: '#/components/schemas/CustomerData'
        CreateCustomerRequest:
            type: object
            properties:
                firstName:
                    type: string
                lastName:
                    type: string
                age:
                    type: integer
                    format: int32
                gender:
                    type: string
                emails:
                    type: array
                    items:
                        $ref: '#/components/schemas/Emaildata'
                phones:
                    type: array
                    items:
                        $ref: '#/components/schemas/Phones'
                addresses:
                    type: array
                    items:
                        $ref: '#/components/schemas/Address'
        CustomerData:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                firstName:
                    type: string
                lastName:
                    type: string
                age:
                    type: integer
                    format: int32
                gender:
                    type: string
                emails:
                    type: array
                    items:
                        $ref: '#/components/schemas/Emaildata'
                phones:
                    type: array
                    items:
                        $ref: '#/components/schemas/Phones'
                addresses:
                    type: array
                    items:
                        $ref: '#/components/schemas/Address'
                createdAt:
                    type: string
                updatedAt:
                    type: string
        DeleteCustomerReply:
            type: object
            properties:
                message:
                    type: string
        Emaildata:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                email:
                    type: string
                type:
                    type: string
                isPrimary:
                    type: boolean
        GetCustomerReply:
            type: object
            properties:
                customer:
                    $ref: '#/components/schemas/CustomerData'
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        ListCustomersReply:
            type: object
            properties:
                customers:
                    type: array
                    items:
                        $ref: '#/components/schemas/CustomerData'
        Phones:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                number:
                    type: string
                type:
                    type: string
                isPrimary:
                    type: boolean
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        UpdateCustomerReply:
            type: object
            properties:
                customer:
                    $ref: '#/components/schemas/CustomerData'
        UpdateCustomerRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                firstName:
                    type: string
                lastName:
                    type: string
                age:
                    type: integer
                    format: int32
                gender:
                    type: string
                emails:
                    type: array
                    items:
                        $ref: '#/components/schemas/Emaildata'
                phones:
                    type: array
                    items:
                        $ref: '#/components/schemas/Phones'
                addresses:
                    type: array
                    items:
                        $ref: '#/components/schemas/Address'
tags:
    - name: Customer